Dynamic Arenaentrance
The entrance to the arena is now marked with an unique armorstand. The rotation of the armorstand is important: The gate is centered on the armorstand with one block to left and one block to the right. It is 2 high. When the gate closes the block above the armorstand is cloned in to all 6 gateblocks.
The armorstand checks the 3x3 in front of him for players with an arenaticket.
Players that are too stupid to leave the arena on their own are teleported one block in front of the arena entrance.